% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxentRaster.R
\name{maxentRaster}
\alias{maxentRaster}
\title{maxentRaster
Generate a raster from a spatial distribution model
fitted with glmnet via maxnet package}
\usage{
maxentRaster(layers, occ, fclasses)
}
\arguments{
\item{layers}{A raster brick of enviromental layers}

\item{occ}{A points data frame}

\item{fclasses}{A vector with feature classes parameters}
}
\value{
A list with a raster and a model fitted with glmnet
}
\examples{
data(iris)
}
