% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomRows.R
\name{randomRows}
\alias{randomRows}
\title{randomRows}
\usage{
randomRows(df, n)
}
\arguments{
\item{df}{A data frame}

\item{n}{A number.}
}
\value{
The \code{n} sampled rows of \code{df}.
}
\description{
Take a sample of the rows of a data frame
}
\examples{
data(iris)
randomRows(iris, 10)
}
